+++ src/apiCalls/configuration/getServerConfiguration.ts
@@ -6,7 +6,7 @@ export async function getServerConfiguration(apiAddress: string) {
   const configUrl = `${apiAddress}/${CONFIG_ENDPOINT}`;
 
   try {
-    const { data } = await axios.get<NetworkType>(configUrl);
+    const { data } = await axiosInstance.get<NetworkType>(configUrl);
     if (data != null) {
       // TODO: egldDenomination will be removed from API when dapp-core v1 will be discontinued
       const egldDenomination = 'egldDenomination';