+++ src/apiCalls/economics/useGetEconomicsInfoFromApi.ts
@@ -1,0 +1,5 @@
+import { useApiFetch } from '../utils';
+import { ECONOMICS_ENDPOINT } from '../endpoints';
+
+export const useGetEconomicsInfoFromApi = () =>
+  useApiFetch({ apiEndpoint: ECONOMICS_ENDPOINT });