+++ src/apiCalls/index.ts
@@ -2,5 +2,5 @@ export * from './transactions';
 export * from './accounts';
 export * from './configuration';
 export * from './economics';
-export * from './getScamAddressData';
+export * from './utils';
 export * from './endpoints';