+++ src/apiCalls/transactions/getTransaction.ts
@@ -1,7 +1,7 @@
-import axios from 'axios';
 import { ServerTransactionType } from 'types/serverTransactions.types';
 import { TRANSACTIONS_ENDPOINT } from '../endpoints';
 import { getTimeout } from './helpers';
+import { axiosInstance } from '../utils';
 
 export interface GetTransactionType {
   apiAddress: string;
@@ -14,7 +14,7 @@ export const getTransaction = ({
   apiAddress,
   apiTimeout
 }: GetTransactionType) => {
-  return axios.get<ServerTransactionType[]>(
+  return axiosInstance.get<ServerTransactionType[]>(
     `${apiAddress}/${TRANSACTIONS_ENDPOINT}/${hash}`,
     {
       ...getTimeout(apiTimeout)