+++ src/apiCalls/transactions/getTransactions.ts
@@ -1,12 +1,12 @@
-import axios from 'axios';
 import { TRANSACTIONS_ENDPOINT } from 'apiCalls/endpoints';
 import { ServerTransactionType } from 'types';
 import { GetTransactionsType } from './getTransactions.types';
 import { getTimeout, getTransactionsParams } from './helpers';
+import { axiosInstance } from '../utils';
 
 export const getTransactions = (props: GetTransactionsType) => {
   const params = getTransactionsParams(props);
-  return axios.get<ServerTransactionType[]>(
+  return axiosInstance.get<ServerTransactionType[]>(
     `${props.apiAddress}/${TRANSACTIONS_ENDPOINT}`,
     {
       params,