+++ src/apiCalls/transactions/sendSignedBatchTransactions.ts
@@ -1,4 +1,3 @@
-import axios from 'axios';
 import { TIMEOUT } from 'constants/network';
 import { buildBatchId } from 'hooks/transactions/helpers/buildBatchId';
 import { addressSelector, networkSelector } from 'reduxStore/selectors';
@@ -10,6 +9,7 @@ import {
   SignedTransactionType
 } from 'types';
 import { TRANSACTIONS_BATCH } from '../endpoints';
+import { axiosInstance } from '../utils';
 
 export interface SendBatchTransactionsPropsType {
   transactions: SignedTransactionType[][];
@@ -49,7 +49,7 @@ export async function sendSignedBatchTransactions({
       id: batchId
     };
 
-    const response = await axios.post<BatchTransactionsResponseType>(
+    const response = await axiosInstance.post<BatchTransactionsResponseType>(
       `${apiAddress}/${TRANSACTIONS_BATCH}`,
       payload,
       {