+++ src/apiCalls/utils/axiosFetcher.ts
@@ -1,0 +1,4 @@
+import { axiosInstance } from 'apiCalls/utils/axiosInstance';
+
+export const axiosFetcher = (url: string) =>
+  axiosInstance.get(url).then((response) => response.data);